requirements.txt
black==25.1.0
click==8.2.1
flake8==7.3.0
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylatexenc==2.10
pytest==8.4.1
PyYAML==6.0.2
typing_extensions==4.15.0
uuid==1.30
